Reorganized components and used mapping input component
