+++ src/DataImportUtility/DataImportUtility.Components/FieldMappingComponents/FieldTransformationComponents/FieldTransformationConfiguration.razor
@@ -1,3 +1,4 @@
+@using DataImportUtility.Components.FieldMappingComponents.ValueTransformationComponents
 @using DataImportUtility.Helpers
 
 @inherits FileImportUtilityComponentBase
