+++ src/DataImportUtility/DataImportUtility.Components/FieldMappingComponents/ValueTransformationComponents/ValueTransformationConfiguration.razor
@@ -1,3 +1,5 @@
+@using DataImportUtility.Components.FieldMappingComponents.FieldTransformationComponents
+
 @inherits FileImportUtilityComponentBase
 
 <div class="value-transform-rule@(DefaultCssClass)"
