+++ src/DataImportUtility/DataImportUtility.Components/FieldMappingComponents/ValueTransformationComponents/ValueTransformationRuleDetails.razor
@@ -1,0 +1,74 @@
+@using DataImportUtility.Components.FieldMappingComponents.FieldTransformationComponents
+@using DataImportUtility.TransformOperations
+
+@inherits FileImportUtilityComponentBase
+
+<div class="value-tranform-details@(DefaultCssClass)"
+     data-from-component="@nameof(FieldTransformationConfiguration)"
+     data-component-instance="@InstanceId"
+     data-mapper-state="@MapperStateId">
+    @switch (ValueTransformation)
+    {
+        case CombineFieldsOperation combineFieldsOperation:
+            @* <CombineFieldsOperationDetails CombineFieldsOperation=combineFieldsOperation /> *@
+            <button type="button" class="@DefaultCssClass">
+                Configure Fields
+            </button>
+
+            <BasicValueTransformationInput ValueTransformation=combineFieldsOperation
+                                           Id=_id
+                                           ApplyDefaultCss=ApplyDefaultCss
+                                           OnAfterInput=HandleOperationDetailChanged />
+            break;
+        case SubstringOperation substringOperation:
+            <SubstringValueTransformationInput ValueTransformation=substringOperation
+                                               Id=_id
+                                               ApplyDefaultCss=ApplyDefaultCss
+                                               OnAfterInput=HandleOperationDetailChanged />
+            break;
+
+        case MapOperation mapOperation:
+            <MapTransformationInput ValueTransformation=mapOperation
+                                    Id=_id
+                                    ApplyDefaultCss=ApplyDefaultCss
+                                    OnAfterInput=HandleOperationDetailChanged />
+
+            break;
+        case RegexMatchOperation regexMatchOperation:
+        case InterpolateOperation interpolateOperation:
+        // If the thing we are doing the interpolation on is a single string, we will only display items relevant to a placeholder for
+        // that string.  If it is a list of strings, we will display the list of strings and the placeholder.
+        default:
+            <BasicValueTransformationInput ValueTransformation=ValueTransformation
+                                           Id=_id
+                                           ApplyDefaultCss=ApplyDefaultCss
+                                           OnAfterInput=HandleOperationDetailChanged />
+            break;
+    }
+</div>
+
+<table class="@(DefaultCssClass)">
+    <thead>
+        <tr>
+            <th>In Value</th>
+            <th>Out Value</th>
+            <th>&nbsp;</th>
+        </tr>
+    </thead>
+    <tbody>
+        <tr>
+            <td>@_myTransformationResult.OriginalValue</td>
+            <td>@_myTransformationResult.Value</td>
+            <td>
+                @if (_myTransformationResult.WasFailure)
+                {
+                    <span title="@_myTransformationResult.ErrorMessage">@Icons.Caution()</span>
+                }
+                else
+                {
+                    <span>&nbsp;</span>
+                }
+            </td>
+        </tr>
+    </tbody>
+</table>
